# Speedup tickstore read date index conversion